requirements.txt
streamlit>=1.31,<1.41
streamlit-drawable-canvas==0.9.3
opencv-python-headless>=4.8
numpy>=1.24
pandas>=2.0
pillow>=10.0
tqdm>=4.66
python-dotenv>=1.0
requests>=2.31
openai>=1.0.0
faiss-cpu>=1.7.4
open-clip-torch>=2.24.0
torch>=2.1
